squelch warnings
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@8956 8aca7d54-2c30-db11-9de9-000461428c89

+++ simtool.cc
@@ -147,7 +147,7 @@ char const *const NOTICE_DISABLED_PUBLIC_WAY = "Not allowed to make publicly own
 size_t strncopy_to_break( char *dest, const char *src, size_t n )
 {
 	const char *p = src;
-	for(  int i=0;  ((uint8)*p)>=32  &&  i<n;  i++  ) { 
+	for(  size_t i=0;  ((uint8)*p)>=32  &&  i<n;  i++  ) {
 		*dest++ = *p++;
 	}
 	*dest = 0;
